Refactor post processing
