+++ modules/gltf/test/lib/gltf-scenegraph-to-tree.spec.js
@@ -1,7 +1,7 @@
 /* eslint-disable max-len, camelcase */
 import test from 'tape-promise/tape';
 
-import getResolvedJson from '@loaders.gl/gltf/lib/gltf-scenegraph-to-tree';
+import {GLTFPostProcessor} from '@loaders.gl/gltf';
 
 const TEST_CASES = [
   {
@@ -10,44 +10,30 @@ const TEST_CASES = [
       json: {
         scenes: [
           {
-            nodes: [0, 1],
+            nodes: [0, 1]
           }
         ],
-        nodes: [
-          {mesh: 0},
-          {mesh: 1}
-        ],
-        meshes: [
-          {},
-          {}
-        ]
+        nodes: [{mesh: 0}, {mesh: 1}],
+        meshes: [{}, {}]
       }
     },
     output: {
       scenes: [
         {
-          nodes: [
-            {mesh: {id: 'mesh-0'}, id: 'node-0'},
-            {mesh: {id: 'mesh-1'}, id: 'node-1'}
-          ],
+          nodes: [{mesh: {id: 'mesh-0'}, id: 'node-0'}, {mesh: {id: 'mesh-1'}, id: 'node-1'}],
           id: 'scene-0'
         }
       ],
-      nodes: [
-        {mesh: {id: 'mesh-0'}, id: 'node-0'},
-        {mesh: {id: 'mesh-1'}, id: 'node-1'}
-      ],
-      meshes: [
-        {id: 'mesh-0'},
-        {id: 'mesh-1'}
-      ]
+      nodes: [{mesh: {id: 'mesh-0'}, id: 'node-0'}, {mesh: {id: 'mesh-1'}, id: 'node-1'}],
+      meshes: [{id: 'mesh-0'}, {id: 'mesh-1'}]
     }
   }
 ];
 
 test('GLTF roundtrip#extensions', t => {
+  const gltfPostProcessor = new GLTFPostProcessor();
   for (const testCase of TEST_CASES) {
-    const json = getResolvedJson(testCase.input);
+    const json = gltfPostProcessor.postProcess(testCase.input);
     t.deepEqual(json, testCase.output, testCase.name);
   }
   t.end();
